+++ src/main/ruby/truffleruby/core/hash.rb
@@ -50,8 +50,13 @@ class Hash
 
   def self.new_from_associate_array(associate_array)
     hash = new
-    associate_array.each do |array|
-      next unless array.respond_to? :to_ary
+    associate_array.each_with_index do |array, i|
+      unless array.respond_to? :to_ary
+        warn "wrong element type #{Truffle::Type.object_class(array)} at #{i} (expected array)"
+        warn 'ignoring wrong elements is deprecated, remove them explicitly'
+        warn 'this causes ArgumentError in the next release'
+        next
+      end
       array = array.to_ary
       unless (1..2).cover? array.size
         raise ArgumentError, "invalid number of elements (#{array.size} for 1..2)"