+++ test/truffle/integration/tracing.sh
@@ -5,5 +5,5 @@ source test/truffle/common.sh.inc
 for f in test/truffle/integration/tracing/*.rb
 do
   echo "$f"
-  jt ruby "$f"
+  jt ruby --no-core-load-path "$f"
 done