+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Bug fixes:
 * Fix `OpenSSL::Digest.{digest,hexdigest,base64digest}` to handle `algorithm, data` arguments (#1889, @bdewater).
 * Fixed `SystemCallError.new` parameter conversion.
 * Fixed `File#{chmod, umask}` argument conversion check.
+* Added warning in `Hash.[]` for non-array elements.
 
 Compatibility: